+++ src/events/IAmDadEvent.ts
@@ -10,12 +10,12 @@ export const IAmDadEvent = new Event<'messageCreate'>({
 
         // Check if string includes 'ik ben' or 'i am'
         if (content.toLowerCase().includes('ik ben')) {
-            const part = content.split('ik ben')[1].trim();
+            const part = content.toLowerCase().split('ik ben')[1].trim();
             const msg = `Hoi ${part}, ik ben Wego Overseer!`;
 
             message.reply(msg);
         } else if (content.toLowerCase().includes('i am')) {
-            const part = content.split('i am')[1].trim();
+            const part = content.toLowerCase().split('i am')[1].trim();
             const msg = `Hi ${part}, I am Wego Overseer!`;
 
             message.reply(msg);